Gcc4 fixes.
darcs-hash:20050908181847-65a35-c9b8b35a36bcd796d545a4dc5f7520f662811619.gz

+++ libsofia-sip-ua/ipt/base64.c
@@ -196,7 +196,7 @@ int base64_e(char buf[], int bsiz, void *data, int dsiz)
   unsigned char *s = (unsigned char *)buf;
   unsigned char *b = (unsigned char *)data;
   unsigned long w;
-  size_t i, n, slack = (size_t)dsiz % 3;
+  size_t i, n, slack = (unsigned)dsiz % 3;
   size_t dsize = (size_t)dsiz - slack, bsize = (size_t)bsiz;
 
   if (bsize == 0)
